backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic==1.12.1",
    "asyncpg==0.28.0",
    "psycopg2-binary>=2.9.11",
    "sqlalchemy[asyncio]>=2.0.46",
    "sqlmodel==0.0.12",
    "pyjwt==2.8.0",
    "python-dotenv==1.0.0",
    "click>=8.3.1",
    "h11>=0.16.0",
    "cryptography>=41.0.0",
    "httpx>=0.25.0",
    "uvicorn>=0.41.0",
    "fastapi[standard]>=0.125.0",
]
